Refactor Convertible into other concerns
Also start of improved parity between Document and Page
